% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_taxonomy.R
\name{add_taxonomy}
\alias{add_taxonomy}
\title{Add BWG-formatted taxonomy to data}
\usage{
add_taxonomy(data_table)
}
\arguments{
\item{data_table}{The input data table, the column 'bwg_name' has to be present}
}
\value{
An updated measurement table that will be used to estimate sizes and gather taxonomy
}
\description{
Add all taxonomy columns needed by the package
}
