% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_measurements.R
\name{append_measurements}
\alias{append_measurements}
\title{Append measurement table}
\usage{
append_measurements(measurement_table, data_table)
}
\arguments{
\item{measurement_table, }{data_table}
}
\value{
An updated measurement table that will be used to estimate sizes
}
\description{
Append measurmement table with size in mm, to be used for size estimation
}
