% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bromeliad_inverts_measurements.R
\name{bromeliad_inverts_measurements}
\alias{bromeliad_inverts_measurements}
\title{Read the \code{bromeliad_inverts_measurements} dataset}
\usage{
bromeliad_inverts_measurements()
}
\value{
A tibble with real body length and body size measurements from bromeliad invertebrates.
It contains columns for taxonomy ("phylum" to "species"), "stage", "abundance", "body_size_mm",
"body_mass_mg", "mass_type" (dry/wet), "data_providers" and "note".
}
\description{
Reads the  \code{bromeliad_inverts_measurements} dataset, a compilation of real invertebrate
body size (mm) and body mass (mg) measurements from bromeliad communities. If you use the
data outside the package, please cite the package.
}
\examples{
# The measurements the package ships with
bromeliad_inverts_measurements()
}
