% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart_path.R
\name{chart_path}
\alias{chart_path}
\title{Summarise path column}
\usage{
chart_path(data_table)
}
\arguments{
\item{data_table}{Output data from hello_metry()}
}
\value{
A list with three elements. Element 1 returns how many times
a raw numerical measurement or biomass was used to estimate (or get
a biomass value). Element 2 returns how many unique estimations
were performed for size categories (weighted average (WA) or size
bins (BIN)) and allometric equations (AE)). Element 3 returns the
level, number of unique estimations, and mean number of values (numerical size measurement or
number of allometric equation) for each of these.
}
\description{
Summarise estimations using path column
}
