% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_measurements.R
\name{combine_measurements}
\alias{combine_measurements}
\title{Combine data with reference measurements}
\usage{
combine_measurements(dats, measurements)
}
\arguments{
\item{dats}{Dataframe to be used for estimation.}

\item{measurements}{Dataframe of reference measurements, or NULL.}
}
\value{
\code{dats} combined with reference measurements, or \code{dats} if
\code{measurements} is NULL.
}
\description{
Combine data and reference measurements into one table, to build the
measurement table from both. Missing abundances in reference measurements
are set to 1.
}
\examples{
dats <-
  data.frame(species = c("sp_a", "sp_b"),
             stage = "larva",
             abundance = c(2, 5),
             size_col = "unknown",
             biomass_col = NA,
             biomass_type = "dry")
measurements <-
  data.frame(species = "sp_a",
             stage = "larva",
             size_col = c(3.1, 2.8),
             biomass_col = c(0.4, 0.3),
             biomass_type = "dry")

combine_measurements(dats, measurements)
}
