% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_data.R
\name{database_data}
\alias{database_data}
\title{Get database data}
\usage{
database_data(estimated)
}
\arguments{
\item{estimated}{Either data without estimation (FALSE) or with estimations (TRUE)}
}
\value{
The same BWG data, either without or with biomass estimated (i.e. calculated with hello_metry(data_table, print = T, biomass = "both", database = F))
}
\description{
Gets BWG data with or without biomass estimates
}
