% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dry_wet.R
\name{dry_wet}
\alias{dry_wet}
\title{Parsing dry and wet measurements}
\usage{
dry_wet(dats, biomass_type)
}
\arguments{
\item{dats}{Input data}

\item{biomass_type}{If biomass should be wet or dry, or most common}
}
\value{
Dataframe with filtered biomass type, and chosen one (dry or wet)
}
\description{
Filter dataframe based on required biomass type (dry or wet).
}
\examples{
# Data holding both kinds of biomass
dats <-
  data.frame(species = c("sp_a", "sp_b"),
             stage = "larva",
             size_col = c(3.1, 2.2),
             biomass_col = c(0.4, 0.2),
             biomass_type = c("dry", "wet"))

# Keep the dry measurements only
dry_wet(dats, biomass_type = "dry")
}
