% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equation_finder.R
\name{equation_finder}
\alias{equation_finder}
\title{Determines if allometric equation is ln or log}
\usage{
equation_finder(size_mm, row, taxo)
}
\arguments{
\item{size_mm}{size in mm}

\item{row}{row of the data frame}

\item{taxo}{taxonomy information for the species, used to get information about the equation}
}
\value{
correct use of equation for one individual of given size
}
\description{
Called when biomass computation happens because some equations are in log10 format
(log10(biomass) = slope x log10(size) + intercept), others in ln format
(ln(biomass) = slope x ln(size) + ln(intercept)). Has an exception argument, so far used for #' Acari, because equation used generates weight in micrograms, not milligrams (argument
controls the conversion)
}
