% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equation_user.R
\name{equation_user}
\alias{equation_user}
\title{Computes biomass with provided equation}
\usage{
equation_user(size_mm, eq, taxo)
}
\arguments{
\item{size_mm}{Input size to be used in equation}

\item{eq}{Allometric equation to be used}

\item{taxo}{Taxonomy information of the species}
}
\value{
Biomass estimation and confidence intervals
}
\description{
Called when biomass computation happens, uses linear modelling
and predict function
}
