% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_check.R
\name{error_check}
\alias{error_check}
\title{Error checking for data frame}
\usage{
error_check(dats, biomass_kind)
}
\arguments{
\item{dats}{The data used to estimate biomass}

\item{biomass_kind}{Should data used in inference be "dry" or "wet", "most_common" or "both"}
}
\value{
Error messages only
}
\description{
Throws custom error message if data not in accepted format
}
