% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firehose.R
\name{firehose}
\alias{firehose}
\title{Determines if allometric equation is ln or log for firfighters}
\usage{
firehose(size_input, row, taxo)
}
\arguments{
\item{size_input}{Input size to be used in equation}

\item{row}{Row of the equation data frame (i.e. one equation)}

\item{taxo}{Taxonomy information of the species}
}
\value{
Correct use of equation for one individual of given size
}
\description{
Called when biomass computation happens because some equations are in log10 format
(log10(biomass) = slope x log10(size) + intercept), others in ln format
(ln(biomass) = slope x ln(size) + ln(intercept)). Has an exception argument, so far used for #' Acari, because equation used generates weight in micrograms, not milligrams (argument
controls the conversion)
}
