% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firetruck.R
\name{firetruck}
\alias{firetruck}
\title{When everything else fails, call the firefighters to bring some literature equations}
\usage{
firetruck(size_input, abundance, taxo)
}
\arguments{
\item{size_input}{Input size to be used in equation}

\item{abundance}{Input abundance to be used in equation}

\item{taxo}{Taxonomy information of the species}
}
\value{
Computed biomass, confidence intervals as strings, and path
}
\description{
Estimates biomass with equations from literature. Handles a number of
equation for each taxonomic class.
}
