% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bwgnames.R
\name{get_bwgnames}
\alias{get_bwgnames}
\title{Get BWG names}
\usage{
get_bwgnames()
}
\arguments{
\item{none}{Directly loads BWG data to get list of BWG names}
}
\value{
Dataframe with BWG names and corresponding taxonomy
}
\description{
Gets dataframe with equation used in estimations
}
