% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_database_data.R
\name{get_database_data}
\alias{get_database_data}
\title{Append measurement table}
\usage{
get_database_data(estimated)
}
\arguments{
\item{estimated}{Either data without estimation (FALSE) or with estimations (TRUE)}
}
\value{
The same BWG data, either without or with biomass estimated (i.e. calculated with hello_metry())
}
\description{
Gets BWG data with or without biomass estimates
}
