% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_equation.R
\name{get_equation}
\alias{get_equation}
\title{Get equation for allometry}
\usage{
get_equation(measurement_table, specname, taxo, level, biomass_kind)
}
\arguments{
\item{measurement_table}{A table with the numerical measurements and biomass used to compute allometric lms}

\item{specname}{BWG name of the species}

\item{taxo}{Taxonomic information about the species}

\item{level}{Taxonomic level of estimation}

\item{biomass_kind}{Should data used in inference be "dry" for just dry biomass, or "both" (default) for both dry and wet biomass.}
}
\value{
List with allometric equations used in estimations, and biomass kind
}
\description{
Once a threshold of three length measurements is reached,
computes a linear model that will be used as allometric equation
}
