% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_equations.R
\name{get_equations}
\alias{get_equations}
\title{Get equations for allometry}
\usage{
get_equations(dats, equation_type, biomass_kind = "both")
}
\arguments{
\item{dats}{A table with the numerical measurements and biomass used to compute allometric models}

\item{equation_type}{Which package should be used to compute the models. LIST TBD}

\item{biomass_kind}{Should data used in inference be "dry" or "wet", "most_common" or "both"}
}
\value{
Dataframe wiht computed models at which level
}
\description{
Computes a linear model that will be used as allometric equation
}
