% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_measurements.R
\name{get_measurements}
\alias{get_measurements}
\title{Get measurement table}
\usage{
get_measurements(database)
}
\arguments{
\item{database}{Should data from the bwg database be used to supplement the measurements (TRUE/FALSE)}
}
\value{
Dataframe with equation used in estimations
}
\description{
Gets dataframe with equation used in estimations
}
