% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hellometry.R
\name{hello_metry}
\alias{hello_metry}
\title{Wrapper function going row-by-row to estimate size (mm) and biomass (mg)}
\usage{
hello_metry(equation_table, measurement_table, data_table, print, biomass_kind)
}
\arguments{
\item{equation_table, }{measurement_table, data_table, print, biomass_kind}
}
\value{
The original data frame,
}
\description{
Wrapper function, returns initial data table with three extra columns: total
biomass from given species, size and abundance, as well as path taken through
the function
}
