% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiply_rows.R
\name{multiply_rows}
\alias{multiply_rows}
\title{Multiply rows based on abundance}
\usage{
multiply_rows(dats)
}
\arguments{
\item{dats}{The data, needs to have a column called abundance}
}
\value{
Original data frame with rows mutiplied based on abundance
}
\description{
In order to have one row per measurement (e.g. if several
individuals had the same size)
}
