% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name_herder.R
\name{name_herder}
\alias{name_herder}
\title{Assign species name if BWG name is missing
Gets or generates mising BWG names}
\usage{
name_herder(data_table)
}
\arguments{
\item{data_table}{The input data table}
}
\value{
The input data table with missing BWG names filled
}
\description{
Looks into measurement table, if finds a single matching taxonomy,
gives this name to the species, otherwise gives a new name. Please
Make sure that each species has a unique taxonomy. If there are two
unidentified species at the same level (e.g. two <i>Stibasoma</i>, then put
any arbitrary string in the 'species' column to contrast between the two)
}
