% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pitilla_data.R
\name{pitilla_data}
\alias{pitilla_data}
\title{Get practice data}
\usage{
pitilla_data(taxonomy = TRUE)
}
\arguments{
\item{taxonomy}{Do you want all taxonomic columns to be included? (TRUE (default)/FALSE)}
}
\value{
The Pitilla data
}
\description{
Gets a toy dataset, collected in Pitilla (Guanacaste, Costa Rica) by Diane Srivastava in 2002.
Some mock rows added by Pierre Rogy to demonstrate some aspects of the package
}
