% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sizest.R
\name{sizest}
\alias{sizest}
\title{Returns size estimates based on taxonomic approach or trait approach}
\usage{
sizest(
  specname,
  size_mm,
  level,
  stage,
  path,
  taxo,
  equation_table,
  measurement_table,
  data_table
)
}
\arguments{
\item{specname, }{size_mm, level, stage, path, taxo, equation_table, measurement_table,
data_table}
}
\value{
extra column with size estimate
}
\description{
Collect abundance/size distribution from species in both measurement and data tables. Look for #' at least three distinct numeric measurements in a given grouping in the following order: species #' to family, traits, suborder to phylum. Once threshold of at least three distinct numeric
measurements is met: if size category “unknown”, return weighted average of all available
measurements, if size category one of S, M, L, divide measurements into three bins of unique
measurement values, if bins not equal, larger bins for S and L. Return average of
measurements within given bin
}
