% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomic_nester.R
\name{taxonomic_nester}
\alias{taxonomic_nester}
\title{Split data to get estimation of size and allometric equations for all levels}
\usage{
taxonomic_nester(
  dats,
  equation_type,
  biomass_kind = "both",
  traits = F,
  cutoff = 3,
  size = TRUE,
  equations = TRUE,
  taxa = c("domain", "kingdom", "phylum", "subphylum", "class", "subclass", "order",
    "suborder", "family", "subfamily", "tribe", "genus", "species", "bwg_name")
)
}
\arguments{
\item{dats}{A table with the numerical measurements and biomass used to compute allometric models}

\item{equation_type}{Which package should be used to compute the models. LIST TBD}

\item{biomass_kind}{Should data used in inference be "dry" or "wet", "most_common" or "both"}

\item{traits}{Custom pathway for traits TBD}

\item{cutoff}{Custom cutoff of the number of measurements for a given level above which the model will be computed}

\item{taxa}{list of taxa at which to compute the equations, must match column names of your data}
}
\value{
Dataframe wiht computed models at which level
}
\description{
Once a custom threshold measurements is reached,
computes a linear model that will be used as allometric equation
}
