% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tipulimo.R
\name{tipulimo}
\alias{tipulimo}
\title{Small function to control for Tipulidae/Limoniidae name change}
\usage{
tipulimo(dats)
}
\arguments{
\item{data}{Dataframe with families as separate entries}
}
\value{
Dataframe with families put together
}
\description{
Gets dataframe with equation used in estimations
}
