% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_taxonomy.R
\name{add_taxonomy}
\alias{add_taxonomy}
\title{Add BWG-formatted taxonomy to data}
\usage{
add_taxonomy(data_table)
}
\arguments{
\item{data_table}{The input data table}

\item{measurement_table}{Table containing all measurements}
}
\value{
An updated measurement table that will be used to estimate sizes and gather taxonomy
}
\description{
Add all columns needed by the package
}
