% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_measurements.R
\name{append_measurements}
\alias{append_measurements}
\title{Append measurement table with new measurements}
\usage{
append_measurements(measurement_table, data_table)
}
\arguments{
\item{measurement_table}{Table containing all measurements}

\item{data_table}{The input data table}
}
\value{
An updated measurement table that will be used to estimate sizes and gather taxonomy
}
\description{
Append measuremement table with new measurements, because this table
also compiles measurements for all species
}
