% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_names.R
\name{append_names}
\alias{append_names}
\title{Append measurement table with new species names}
\usage{
append_names(measurement_table, dats, level_list, nothing)
}
\arguments{
\item{measurement_table}{Table containing all measurements}

\item{dats}{Dataframe to be used for estimations}

\item{level_list}{List of taxonomic levels to be used in the measurement table}
}
\value{
An updated measurement table that will be used to estimate sizes and
gather taxonomy
}
\description{
Append measuremement table with new species names, because this table
also compiles taxonomy for all species
}
