% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_names.R
\name{append_names}
\alias{append_names}
\title{Append measurement table with new species names}
\usage{
append_names(measurement_table, data_table)
}
\arguments{
\item{measurement_table}{Table containing all measurements}

\item{data_table}{The input data table}
}
\value{
An updated measurement table that will be used to estimate sizes and gather taxonomy
}
\description{
Append measuremement table with new species and measurements, because this table
also compiles taxonomy for all species
}
