% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart_path.R
\name{chart_path}
\alias{chart_path}
\title{Summarise path column}
\usage{
chart_path(data_table)
}
\arguments{
\item{data_table}{Output data from hello_metry()}
}
\value{
A list with two elements. Element 1 returns how many unique
estimations were performed for size categories (weighted average (WA) or size
bins (BIN)) and allometric equations (AE)), and how many biomass values had raw wet or dry
equivalence in the database. Element 2 returns the level, number of
unique estimations, and mean number of values (numerical size measurement or
number of allometric equation) for each of these.
}
\description{
Summarise estimations using path column
}
