% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_checker.R
\name{data_checker}
\alias{data_checker}
\title{Set of checks to perform on the data to make sure it can be handled by \code{hellometry()}}
\usage{
data_checker(dats, biomass_type)
}
\arguments{
\item{dats}{Dataframe to be used for estimations}

\item{biomass_type}{"dry"/"wet". Whether the biomass used in inference should
be dry (default) or wet. See \code{dry_wet()} for details.}
}
\description{
Will fail if any condition not met
}
