% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_checker.R
\name{data_checker}
\alias{data_checker}
\title{Set of checks to perform on the data to make sure it can be handled by \code{hellometry()}}
\usage{
data_checker(dats, biomass_type, database, nothing)
}
\arguments{
\item{dats}{Dataframe to be used for estimations}
}
\description{
Will fail if any condition not met
}
