% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_data.R
\name{database_data}
\alias{database_data}
\title{Get database data}
\usage{
database_data()
}
\arguments{
\item{none}{will just read the local copy of the data}
}
\value{
Latest version of BWG data on file
}
\description{
Gets BWG data without biomass estimates
}
