% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dry_wet.R
\name{dry_wet}
\alias{dry_wet}
\title{Parsing dry and wet measurements}
\usage{
dry_wet(dats, biomass_kind)
}
\arguments{
\item{dats}{Input data}

\item{biomass_kind}{If biomass should be wet or dry}
}
\value{
Dataframe with filtered biomass kind, and chosen one (dry or wet)
}
\description{
Filter dataframe based on required biomass type (dry or wet),
or most common one (both)
}
