% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dry_wet.R
\name{dry_wet}
\alias{dry_wet}
\title{Parsing dry and wet measurements}
\usage{
dry_wet(dats, biomass_type)
}
\arguments{
\item{dats}{Input data}

\item{biomass_type}{If biomass should be wet or dry, or most common}
}
\value{
Dataframe with filtered biomass type, and chosen one (dry or wet)
}
\description{
Filter dataframe based on required biomass type (dry or wet).
}
