% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/full_estimation_table.R
\name{full_estimation_table}
\alias{full_estimation_table}
\title{Returns size and biomass estimations for every taxon in measurement table}
\usage{
full_estimation_table(level_list, measurement_table, what)
}
\arguments{
\item{level_list}{Vector of taxonomic levels over which to iterate estimations}

\item{measurement_table}{A table with the numerical measurements and biomass
used to compute allometric lms}

\item{what}{Should size (what = "size_mm") or biomass (what = "biomass_mg") be
estimated?}
}
\value{
Tibble with size estimations or allometric models used in estimations
}
\description{
Returns size and biomass estimations for every taxon in measurement table
}
