% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_biomass.R
\name{get_biomass}
\alias{get_biomass}
\title{Estimates biomass (mg)}
\usage{
get_biomass(
  specname,
  level,
  size_input,
  abundance,
  stage,
  path,
  taxo,
  measurement_table,
  biomass_kind
)
}
\arguments{
\item{specname}{BWG name of the species}

\item{level}{Taxonomic level of estimation}

\item{abundance}{Abundance of individuals of the species/size combination}

\item{stage}{Developmental stage of the invertebrate (larva/pupa/adult)}

\item{path}{Path taken trough function}

\item{taxo}{Taxonomic information about the species}

\item{measurement_table}{Table containing all measurements}

\item{biomass_kind}{Should data used in inference be "dry" for just dry biomass, or "both" (default) for both dry and wet biomass.
If both (the default) is chosen, then the function will determine which dry or wet equations or raw weight is present, and choose
the most numerous. If there is the same number of dry and wet equations, dry equations are always favoured. Dry and wet
measurements are never mixed.}

\item{size_mm}{Size in mm}

\item{equation_table}{Table containing all allometric equation}
}
\value{
new column with estimate biomass value
}
\description{
Looks if raw biomass value exists for given measurement, use it if #' available.
Looks for at least three length measurement to compute allometric equation in a given
grouping in the following order: species to family, traits,
suborder to class. Once threshold of at least three length measurements
equation is met calls the get_equation function to compute equation
}
