% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_biomass.R
\name{get_biomass}
\alias{get_biomass}
\title{Estimates biomass (mg)}
\usage{
get_biomass(
  specname,
  level,
  size_mm,
  abundance,
  stage,
  path,
  taxo,
  equation_table,
  measurement_table,
  biomass_kind
)
}
\arguments{
\item{specname, }{level, size_mm, abundance, stage, path, taxo, equation_table, #' measurement_table, biomass_kind}
}
\value{
new column with estimate biomass value
}
\description{
Look if raw biomass value exists for given measurement, use it if #' available. Look for at least one allometric equation in a given #' grouping in the following order: species to family, traits,
suborder to phylum. Once threshold of at least one allometric
equation is met: if there is only one equation, compute value of #' biomass; if there is more than one allometric equation, average #' the computed values of biomass.
}
