% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_equations.R
\name{get_equations}
\alias{get_equations}
\title{Get equation table}
\usage{
get_equations()
}
\value{
Dataframe with equation used in estimations
}
\description{
Gets dataframe with equation used in estimations
}
