% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_equations.R
\name{get_equations}
\alias{get_equations}
\title{Get equations for allometry}
\usage{
get_equations(dats, equation_type = "lm")
}
\arguments{
\item{dats}{A table with the numerical measurements and biomass used to compute allometric models}

\item{equation_type}{Which package should be used to compute the models. "lm" is default, "ols"
is for lmodel2 (options used are "interval" for both x and y, and 100 permutations)}
}
\value{
Dataframe with computed models at which level, \emph{please double check models}
}
\description{
Computes a linear model that will be used as allometric equation
}
