% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/%notin%.R
\name{\%notin\%}
\alias{\%notin\%}
\title{Negate \%in\%}
\usage{
... \%notin\% NA
}
\arguments{
\item{x}{regular \%in\% parameter}

\item{y}{regular \%in\% parameter}
}
\value{
Input without supplied parameters
}
\description{
Just a lot more practical
}
