% Generated by roxygen2: do not edit by hand
<<<<<<< HEAD
% Please edit documentation in R/operators.R
\name{\%notin\%}
\alias{\%notin\%}
\title{Not In Operator}
\usage{
x \%notin\% table
}
\arguments{
\item{x}{vector or NULL: the values to be matched.}

\item{table}{vector or NULL: the values to be matched against.}
}
\value{
A logical vector of the same length as \code{x}.
}
\description{
Returns logical vector indicating if elements are \strong{not} in a given set.
}
\examples{
1 \%notin\% c(2, 3, 4) # TRUE
"a" \%notin\% letters # FALSE
=======
% Please edit documentation in R/%notin%.R
\name{\%notin\%}
\alias{\%notin\%}
\title{Negate \%in\%}
\usage{
... \%notin\% NA
}
\arguments{
\item{x}{regular \%in\% parameter}

\item{y}{regular \%in\% parameter}
}
\value{
Input without supplied parameters
}
\description{
Just a lot more practical
>>>>>>> main
}
