% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{\%notin\%}
\alias{\%notin\%}
\title{Not In Operator}
\usage{
x \%notin\% table
}
\arguments{
\item{x}{vector or NULL: the values to be matched.}

\item{table}{vector or NULL: the values to be matched against.}
}
\value{
A logical vector of the same length as \code{x}.
}
\description{
Returns logical vector indicating if elements are \strong{not} in a given set.
}
