% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello_metry.R
\name{hello_metry}
\alias{hello_metry}
\title{Wrapper function going row-by-row to estimate size (mm) and biomass (mg)}
\usage{
hello_metry(data_table, print, biomass_kind, database)
}
\arguments{
\item{data_table, }{print, biomass_kind, database}
}
\value{
The original data frame,
}
\description{
Wrapper function, returns initial data table with three extra columns: total
biomass from given species, size and abundance, as well as path taken through
the function
}
