% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello_metry.R
\name{hello_metry}
\alias{hello_metry}
\title{Wrapper function to estimate size (mm) and biomass (mg) from an input dataset}
\usage{
hello_metry(
  dats,
  level_list,
  biomass_type = "dry",
  database = TRUE,
  nothing = FALSE
)
}
\arguments{
\item{dats}{The input data table, please include columns columns "abundance",
"size_mm", "biomass_mg", "stage" (larva/pupa/adult, please only put 'adult' for
adult insects)}

\item{level_list}{A character vector indicating the taxonomic levels present
in your data, in indecreasing order of taxonomic resolution.}

\item{biomass_type}{"dry"/"wet". Should data used in inference be "dry" for
just dry biomass (default), or "wet" for just wet biomass. See \code{dry_wet()}
for more information on how this works.}

\item{database}{Logical. Should data from the BWG database be used to supplement
a set of BWG-specific measurements, not present in the database
(TRUE(default)/FALSE). This function uses numerical measurements both from the
BWG database and from the data you provide. Only put FALSE if you are doing
estimations from data already present in the BWG database (estimation for
these data available with \code{database_data(TRUE)}).}

\item{nothing}{Logical. If TRUE, no BWG-specific data will be used for estimations,
only the data you provide. This is useful if you have your own measurement
database or work on a different system (default FALSE).}
}
\value{
A list with three dataframes:
\itemize{
\item data: the input data with added size and biomass estimations, and new columns
with the taxonomic level and name of the taxon at which the estimation
was made.
\item size_estimations: dataframe with size estimations that were used,
with columns for taxonomic level, name and size category
of the estimation.
\item model_estimations: dataframe with biomass models that were used, with columns
for taxonomic level and name.
See \code{full_estimation_table()} to get all possible size estimations and models
for your data
}
}
\description{
Wrapper function, input your data and get size and measurement .
Please name column with number of specimen "abundance", column with measurement
in mm "size_mm", column with life stage (larva/pupa/adult) "stage", column
with biomass type (dry/wet) "biomass_type", and the column with biomass in mg
"biomass_mg".
If you do not have a numerical measurement for a given specimen, the function
can do size estimations if you input "small", "medium", "large" or "unknown".
In this case, the algorithm will use existing size measurements for the species,
and use the size distribution to estimate "small", "medium" and "large" inputs,
or a weighted average of all measurements if the input is "unknown". For biomass
measurements, just leave NA for those cells you want estimations in.
If you are using BWG data, please make sure that the column with BWG names is
called "bwg_name".
}
\details{
Please make sure that the level_list parameter has levels in increasing order
of resolution, e.g. from species to order.
}
