% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello_metry.R
\name{hello_metry}
\alias{hello_metry}
\title{Wrapper function going row-by-row to estimate size (mm) and biomass (mg)}
\usage{
hello_metry(data_table, print = FALSE, biomass_kind = "both", database = TRUE)
}
\arguments{
\item{data_table}{The input data table, please include columns columns "abundance", "bwg_name", "size_mm", "stage" (larva/pupa/adult, please only put 'adult' for adult insects) are present}

\item{biomass_kind}{Should data used in inference be "dry" for just dry biomass, or "both" (default) for both dry and wet biomass.
If both (the default) is chosen, then the function will determine which dry or wet equations or raw weight is present, and choose
the most numerous. If there is the same number of dry and wet equations, dry equations are always favoured. Dry and wet
measurements are never mixed.}

\item{database}{Should data from the bwg database be used to supplement the measurements (TRUE(default)/FALSE). This function uses numerical
measurements both from the BWG database and from the data you provide. Only put FALSE if you are doing estimations from
data already present in the BWG database (estimation for these data available with database_data(TRUE)).}
}
\value{
Your initial data table with three extra columns: value (in mm)
of size estimation (NA if not done), total biomass for given row (in mg),
as well as path taken through the function. The nomenclature of this path column is as follows.
\subsection{Special cases}{
\itemize{
\item "null_biomass": abundance was 0
\item "_size_estimaton_failed": there were not enough close relatives or species with matching traits to compute size
\item "_biomass_estimation_failed": there were no allometric equation at to estimate biomass
\item "_raw_dry/wet": if there was a raw (direct) measurement for that particular species-size combination, and whether that measurement is dry or wet biomass
\item "_external_dry_equations": data provided was not sufficient so equations from the literature were used
}
}

\subsection{Regular cases}{
\itemize{
\item if the species goes through estimation of size (sizest()):
estimation kind:level_number of measurements.
For example,  WA:genus_5 (size estimation using weighted average on 5 measurements from the species' genus), BIN:subfamily_3 (size estimation using size bins (S, M, L) on three measurement a the subfamily level)
\item if the allometric equations are used (get_biomass()):
biomass estimation:taxonomic level of inference_number of equations_dry/wet.
For example, -AE:bwg_name_wet (allometric equation from wet biomass at the species level), -AE:subclass_dry (allometric equation from dry biomass at the subclass level)
}If a species went through both size estimation and biomass estimations, the path will be composite, e.g. WA:genus_5-AE:bwg_name_wet
}
}
\description{
Wrapper function, input your data and get size and measurement . Please name column with number of specimen "abundance",
column with BWG name "bwg_name",  and column with measurement in mm "size_mm".
If you do not have a numerical measurement for a given specimen, the algorithm can
do size estimations if you input "small", "medium", "large" or "unknown". In this
case, the algorithm will use existing size measurements for the species, and use
the size distribution to estimate "small", "medium" and "large" inputs, or a weighted
average of all measurements if the input is "unknown".
}
