% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello_metry.R
\name{hello_metry}
\alias{hello_metry}
\title{Wrapper function going row-by-row to estimate size (mm) and biomass (mg)}
\usage{
hello_metry(data_table, print, biomass_kind, database)
}
\arguments{
\item{data_table}{The input data table print, make sure columns "abundance", "bwg_name", "size_mm" are present}

\item{print}{Do you want to see the printing of the rows (TRUE/FALSE)}

\item{biomass_kind}{"dry" for just dry biomass, or "both" for both dry and wet}

\item{database}{Should data from the bwg database be used to supplement the measurements (TRUE/FALSE)}
}
\description{
Wrapper function, returns initial data table with three extra columns: total
biomass from given species, size and abundance, as well as path taken through
the function
}
