% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hellometry-package.R
\docType{package}
\name{hellometry-package}
\alias{hellometry-package}
\title{hellometry: Biomass estimations for the BWG database and beyond}
\description{
Estimates body size (mm) and biomass (mg) of bromeliad invertebrates from allometric relationships. Although developed for the Bromeliad Working Group (BWG) database, it works with any data that follow a few column-naming conventions.
}
\author{
\strong{Maintainer}: Pierre Rogy \email{pierre.rogy@gmail.com} (\href{https://orcid.org/0000-0002-3002-0059}{ORCID})

}
\keyword{internal}
