% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_trait_table.R
\name{make_trait_table}
\alias{make_trait_table}
\title{Prepare data to do estimations with traits in \code{full_estimation_table}}
\usage{
make_trait_table(measurement_table, trait_columns)
}
\arguments{
\item{measurement_table}{A table with the numerical measurements and biomass
used to compute allometric lms}

\item{trait_columns}{List of traits to match, should be column names in measurement_table}
}
\value{
Tibble of four columns: level = "traits", group_focus_species, stage of focus species,
bwg_name of matched species
}
\description{
Remove columns with taxonomy, and replace them with trait-based groupings
}
