% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_trait_table.R
\name{make_trait_table}
\alias{make_trait_table}
\title{Modify measurement table so that estimations are done with traits}
\usage{
make_trait_table(measurement_table, level_list, trait_columns)
}
\arguments{
\item{level_list}{list of taxonomic levels in measurement_table}

\item{trait_columns}{List of traits to match}

\item{mmeasurement_table}{A table with the numerical measurements and biomass
used to compute allometric lms}
}
\value{
Tibble of four columns: level = "traits", group_focus_species, stage of focus species,
bwg_name of matched species
}
\description{
Remove columns with taxonomy, and replace them with trait-based groupings
}
