% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matcher_of_traits.R
\name{matcher_of_traits}
\alias{matcher_of_traits}
\title{Finds which species have matching traits}
\usage{
matcher_of_traits(specname, measurement_table)
}
\arguments{
\item{measurement_table}{Table containing all measurements}

\item{species_name}{BWG name of the species}
}
\value{
List of species with matching traits
}
\description{
Called if grouping in sizest or get_biomass is “traits”, i.e. if threshold of at
least three distinct measurements or one allometric equation not met at family level. Match
traits of maximum body size, body shape, locomotion and morphological defense across all species #' in the database. Those species with +/- 1 match are added to a species pool to get allometric
equations or size estimates (instead of a taxonomic grouping). Does not proceed if any NA is
found in the trait values of the species
}
