% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matcher_of_traits.R
\name{matcher_of_traits}
\alias{matcher_of_traits}
\title{Finds which species have matching traits}
\usage{
matcher_of_traits(measurement_table, trait_columns)
}
\arguments{
\item{measurement_table}{A table with the numerical measurements and biomass
used to compute allometric lms}

\item{trait_columns}{List of traits to match, should be column names in measurement_table}
}
\value{
Tibble of four columns: level = "traits", group_focus_species, stage of focus species,
bwg_name of matched species
}
\description{
Evaluates all unique trait groupings and returns species that have the same
traits +/- 1
}
