% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_squared.R
\name{r_squared}
\alias{r_squared}
\title{Get the R2 of a fitted allometric model}
\usage{
r_squared(model)
}
\arguments{
\item{model}{A fitted allometric model}
}
\value{
A single R2 value, NA if none could be computed
}
\description{
Wraps \code{performance::r2()}, which returns the R2 suited to the model it is
given, so that any kind of model can be filtered the same way in
\code{full_estimation_table()}
}
