% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sizest.R
\name{sizest}
\alias{sizest}
\title{Get all size estimates}
\usage{
sizest(dats)
}
\arguments{
\item{dats}{Input data}
}
\value{
A vector with weighted average of sieze, and estimations for small, medium and large sizes
}
\description{
Computes weigthed average and small, medium, large size categories for input data
}
