% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trini_communities.R
\name{trini_communities}
\alias{trini_communities}
\title{Read the \code{trini_communities} dataset}
\usage{
trini_communities()
}
\value{
A tibble with one row per taxon per bromeliad, with columns for
taxonomy, abundance, and bromeliad ID.
}
\description{
Reads the \code{trini_communities} dataset, a set of invertebrate
community samples from 20 Trinidadian bromeliads (from @Rogy2024). It is provided
as a toy dataset to try out the package, alongside \code{\link[=bromeliad_inverts_measurements]{bromeliad_inverts_measurements()}}.
}
